requirements.txt
# AI Interviewer - Python Dependencies

# Web Framework
Flask>=3.0.0
flask-cors>=4.0.0

# AI & LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0

# Audio Processing
faster-whisper>=0.10.0
pyaudio>=0.2.13
soundfile>=0.12.1

# Text-to-Speech (optional - install manually if needed)
# kokoro-tts - May require manual installation from source

# Data Processing & Validation
pydantic>=2.5.0
PyYAML>=6.0.1

# Environment & Configuration
python-dotenv>=1.0.0

# Note: Standard library modules used (no installation needed):
# datetime, os, json, re, typing, tempfile, wave, pathlib
